refactor(anthropic): extract AnthropicModelFactory into plain embabel-agent-anthropic module for starter-free BYOK (#1752)
