+++ embabel-agent-dependencies/pom.xml
@@ -223,6 +223,12 @@
 
             <dependency>
                 <groupId>com.embabel.agent</groupId>
+                <artifactId>embabel-agent-anthropic</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+
+            <dependency>
+                <groupId>com.embabel.agent</groupId>
                 <artifactId>embabel-agent-onnx</artifactId>
                 <version>${project.version}</version>
             </dependency>
