+++ pom.xml
@@ -38,6 +38,7 @@
         <module>embabel-agent-starters</module>
         <module>embabel-agent-test-support</module>
         <module>embabel-agent-openai</module>
+        <module>embabel-agent-anthropic</module>
         <module>embabel-agent-onnx</module>
     </modules>
